reorganized code and started readme
