+++ README.md
@@ -1,1 +1,38 @@
-## Find-My-Employee
+# FIND-MY-EMPLOYEE
+
+A mySQL Employee Tracker powered by:
+
+- mySQL2
+- inquirer
+- node.js
+- dotenv
+
+[This is an external link to the 'Find My Employee' repository](https://github.com/rhodemc/find-my-employee)
+
+## Description
+
+<!-- I was motivated to create a logo generator because it is an efficient way to make a quick company logo without distracting from the coding work, itself.
+
+When using this logo generator, you'll find responsive prompts that create a brilliant logo in SVG file format. -->
+
+## Installation
+
+Once the repository is downloaded, please navigate to the correct file path in your command terminal and enter 'npm i' to install the necessary dependencies.
+
+## Usage and Testing
+
+Once necessary dependencies are installed, run 'node index.js' in your command terminal to begin the application. Follow the prompts to add, view, and update employees, roles, and departments.
+
+Future updates will include more customizable data.
+
+![Find My Employee - A mySQL Employee Tracker](images/find-my-employee.png)
+
+<!-- [This is an external link to the 'Find My Employee' usage video]() -->
+
+## Credits
+
+N/A
+
+## License
+
+This project is licensed by MIT. Please refer to the documentation in the repo.
